+++ apollo-adminservice/src/main/java/com/ctrip/framework/apollo/adminservice/controller/AppController.java
@@ -16,15 +16,22 @@
  */
 package com.ctrip.framework.apollo.adminservice.controller;
 
+import com.ctrip.framework.apollo.biz.config.BizConfig;
+import com.ctrip.framework.apollo.biz.entity.AccessKey;
+import com.ctrip.framework.apollo.biz.service.AccessKeyService;
 import com.ctrip.framework.apollo.biz.service.AdminService;
 import com.ctrip.framework.apollo.biz.service.AppService;
+import com.ctrip.framework.apollo.common.constants.AccessKeyMode;
 import com.ctrip.framework.apollo.common.dto.AppDTO;
 import com.ctrip.framework.apollo.common.entity.App;
 import com.ctrip.framework.apollo.common.exception.BadRequestException;
 import com.ctrip.framework.apollo.common.exception.NotFoundException;
 import com.ctrip.framework.apollo.common.utils.BeanUtils;
+import com.ctrip.framework.apollo.common.utils.UniqueKeyGenerator;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
 import org.springframework.data.domain.Pageable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -41,12 +48,19 @@ import java.util.Objects;
 @RestController
 public class AppController {
 
+  private static final Logger logger = LoggerFactory.getLogger(AppController.class);
+
   private final AppService appService;
   private final AdminService adminService;
+  private final AccessKeyService accessKeyService;
+  private final BizConfig bizConfig;
 
-  public AppController(final AppService appService, final AdminService adminService) {
+  public AppController(final AppService appService, final AdminService adminService,
+      final AccessKeyService accessKeyService, final BizConfig bizConfig) {
     this.appService = appService;
     this.adminService = adminService;
+    this.accessKeyService = accessKeyService;
+    this.bizConfig = bizConfig;
   }
 
   @PostMapping("/apps")
@@ -58,8 +72,32 @@ public class AppController {
     }
 
     entity = adminService.createNewApp(entity);
+    if (bizConfig.isAccessKeyAutoProvisionEnabled()) {
+      try {
+        accessKeyService.create(entity.getAppId(), createDefaultAccessKey(entity));
+      } catch (Exception ex) {
+        logger.warn("Failed to auto-provision access key for appId={}", entity.getAppId(), ex);
+      }
+    }
 
     return BeanUtils.transform(AppDTO.class, entity);
+  }
+
+  private AccessKey createDefaultAccessKey(App app) {
+    String operator = app.getDataChangeCreatedBy();
+    if (StringUtils.isBlank(operator)) {
+      operator = app.getDataChangeLastModifiedBy();
+    }
+    if (StringUtils.isBlank(operator)) {
+      operator = app.getOwnerName();
+    }
+    AccessKey accessKey = new AccessKey();
+    accessKey.setAppId(app.getAppId());
+    accessKey.setSecret(UniqueKeyGenerator.generateId());
+    accessKey.setMode(AccessKeyMode.FILTER);
+    accessKey.setEnabled(true);
+    accessKey.setDataChangeCreatedBy(operator);
+    return accessKey;
   }
 
   @DeleteMapping("/apps/{appId:.+}")